Refactor - moved HumanSoldierToken to shared area
