+++ Mage.Sets/src/mage/sets/shadowsoverinnistrad/UlvenwaldMysteries.java
@@ -46,6 +46,7 @@ import mage.filter.predicate.permanent.TokenPredicate;
 import mage.game.Game;
 import mage.game.events.GameEvent;
 import mage.game.events.GameEvent.EventType;
+import mage.game.permanent.token.HumanSoldierToken;
 
 /**
  *
